requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
openpyxl==3.1.2